The Joiful CSS Update
This is my meager contribution thus far. I overhauled the CSS to make it more cohesive for all pages, especially given that the headers and footers are added via PHP. This is rough and based off of one night's work, but definitely better than before (methinks). I'll keep working and refining as need be.

There is a corresponding message in the Discord chat.